requirements.txt
redis
wrapt
pytest
virtualenv
tox
python-dateutil
